+++ include/GridLib/Profile.hpp
@@ -13,10 +13,10 @@
 
 namespace GridLib
 {
-    double get_value(const GridView& grid, const Xyz::Vector2D& pos);
+    float get_value(const GridView& grid, const Xyz::Vector2D& pos);
 
     [[nodiscard]] std::vector<Xyz::Vector3D>
-    make_profile(Chorasmia::ArrayView2D<double> grid,
+    make_profile(Chorasmia::ArrayView2D<float> grid,
                  const Xyz::Vector2D& from,
                  const Xyz::Vector2D& to,
                  size_t steps);
