+++ tests/test_Profile.cpp
@@ -25,10 +25,10 @@ TEST_CASE("Test get_value")
 
 TEST_CASE("Profile without clipping")
 {
-    Chorasmia::Array2D<double> grid({1, 2, 3,
-                                     4, 5, 6,
-                                     7, 8, 9},
-                                    3, 3);
+    Chorasmia::Array2D<float> grid({1, 2, 3,
+                                    4, 5, 6,
+                                    7, 8, 9},
+                                   3, 3);
     auto contour = GridLib::make_profile(grid.view(), {0.5, 0.5}, {1.5, 1.5}, 2);
     REQUIRE(contour.size() == 3);
     REQUIRE_THAT(contour[0][2], WithinAbs(3, 1e-12));
@@ -38,10 +38,10 @@ TEST_CASE("Profile without clipping")
 
 TEST_CASE("Profile with clipping")
 {
-    Chorasmia::Array2D<double> grid({1, 2, 3,
-                                     4, 5, 6,
-                                     7, 8, 9},
-                                    3, 3);
+    Chorasmia::Array2D<float> grid({1, 2, 3,
+                                    4, 5, 6,
+                                    7, 8, 9},
+                                   3, 3);
     auto contour = GridLib::make_profile(grid.view(), {4.5, 3}, {-2.5, -0.5}, 7);
     REQUIRE(contour.size() == 4);
     REQUIRE_THAT(contour[0][2], WithinAbs(8.75, 1e-12));
@@ -52,10 +52,10 @@ TEST_CASE("Profile with clipping")
 
 TEST_CASE("Profile with transformation")
 {
-    Chorasmia::Array2D<double> array({1, 2, 3,
-                                      4, 5, 6,
-                                      7, 8, 9},
-                                     3, 3);
+    Chorasmia::Array2D<float> array({1, 2, 3,
+                                     4, 5, 6,
+                                     7, 8, 9},
+                                    3, 3);
     GridLib::Grid grid(array);
     grid.set_planar_coords(GridLib::PlanarCoords{500.0, 1000.0});
     grid.set_row_axis({{10, 0, 0}, GridLib::Unit::METERS});
